Changed intrinsic value to only use references and restructured main modules into seperate files
